requirements.txt
Flask
gunicorn
Jinja2
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp36-cp36m-linux_x86_64.whl
res-facenet
pillow
torch-mtcnn

click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
#Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
#Flask-MQTT==1.1.1
#Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
#paho-mqtt==1.1
protobuf==3.12.2
#psycopg2==2.8.5
python-etcd==0.4.5
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3

